# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "用户中心服务 API 文档",
        "title": "User Hub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/v1/user-hub/wechat/login": {
            "post": {
                "description": "用户通过提供微信小程序 wx.login() 获取的 code，进行登录或（如果首次登录）自动注册账户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "微信小程序认证"
                ],
                "summary": "微信小程序登录或注册",
                "parameters": [
                    {
                        "description": "包含微信小程序 code 的请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WechatMiniProgramLoginData"
                        }
                    },
                    {
                        "enum": [
                            "web",
                            "wechat",
                            "app"
                        ],
                        "type": "string",
                        "default": "wechat",
                        "description": "客户端平台类型",
                        "name": "X-Platform",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录或注册成功，返回用户信息及访问和刷新令牌",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPILoginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如JSON格式错误、code为空、平台类型无效) 或 业务逻辑错误 (如微信 code 无效或已过期、用户状态异常)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如调用微信API失败、数据库操作失败、令牌生成失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "docs.SwaggerAPIErrorResponseString": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "type": "string"
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "docs.SwaggerAPILoginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.LoginResponse"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.WechatMiniProgramLoginData": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "description": "Code 微信小程序通过 wx.login() 获取的临时授权码\n- 必填，用于后端换取 openid 和 session_key",
                    "type": "string"
                }
            }
        },
...
        "vo.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token 对",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.TokenPair"
                        }
                    ]
                },
                "userManage": {
                    "description": "用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.Userinfo"
                        }
                    ]
                }
            }
        },
...
        "vo.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "新认证令牌",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "新刷新令牌（可选）",
                    "type": "string"
                }
            }
        },
...
        "vo.Userinfo": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string"
                }
            }
        }
    }
}